backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
apscheduler==3.10.4
python-telegram-bot==21.3
web3==6.20.1
gunicorn==22.0.0
